Version 1
This is the first class design diagrams of the metric to imperial conversion program. 

Note: 
Will make changes if necessary. We will see after the coming presentation of our design going forward if there is any
means to improve upon its coherence. The team will go back and discuss redesigning and brainstorm ideas.